+++ src/sgfc/message/SgfcMessage.h
@@ -46,8 +46,8 @@ namespace LibSgfcPlusPlus
     SgfcMessage(
       SgfcMessageID messageID,
       SgfcMessageType messageType,
-      int lineNumber,
-      int columnNumber,
+      unsigned long lineNumber,
+      unsigned long columnNumber,
       bool isCriticalMessage,
       int libraryErrorNumber,
       const std::string& messageText,
@@ -73,8 +73,8 @@ namespace LibSgfcPlusPlus
 
     virtual SgfcMessageType GetMessageType() const override;
 
-    virtual int GetLineNumber() const override;
-    virtual int GetColumnNumber() const override;
+    virtual unsigned long GetLineNumber() const override;
+    virtual unsigned long GetColumnNumber() const override;
 
     virtual bool IsCriticalMessage() const override;
 
@@ -86,8 +86,8 @@ namespace LibSgfcPlusPlus
   private:
     SgfcMessageID messageID;
     SgfcMessageType messageType;
-    int lineNumber;
-    int columnNumber;
+    unsigned long lineNumber;
+    unsigned long columnNumber;
     bool isCriticalMessage;
     int libraryErrorNumber;
     std::string messageText;
